pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zembeddings"
version = "0.1.0"
description = "Trajectory analysis in semantic embedding space — velocity, acceleration, jerk, Kalman-filter boundary detection, and fixation/return classification."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # ── Core ──────────────────────────────────────────────────────────
    "openai>=1.0",
    "tiktoken>=0.5",
    "numpy>=1.24",
    "scipy>=1.11",
    # ── Config & IO ──────────────────────────────────────────────────
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    # ── Data ─────────────────────────────────────────────────────────
    "pandas>=2.0",
    # ── Visualization ────────────────────────────────────────────────
    "matplotlib>=3.7",
    # ── Database (pgvector) ──────────────────────────────────────────
    "psycopg2-binary>=2.9",
    "pgvector>=0.2",
    # ── Dimensionality reduction ─────────────────────────────────────
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ipython",
    "jupyter",
]

[project.scripts]
zembeddings-run = "scripts.run_experiment:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
